chore: refactor settings modal with multi comps
