+++ components/BibleVersionSettings.vue
@@ -1,0 +1,94 @@
+<template>
+  <div>
+    <div
+      v-for="bibleVersion in bibleVersionOptions"
+      :key="bibleVersion"
+      class="bible-version-card relative pb-4 mb-4 border-b border-gray-200 last:border-0 dark:border-gray-700 flex items-center justify-between gap-4"
+    >
+      <UProgress
+        class="absolute inset-0 top-auto rounded-none opacity-0"
+        :class="{
+          'opacity-1': bibleVersionLoading === bibleVersion?.id,
+        }"
+        :value="parseInt(bibleDownloadProgress)"
+        :max="100"
+        size="xs"
+      />
+      <div class="col">
+        <div class="text-md">{{ bibleVersion?.id }}</div>
+        <div class="text-sm text-gray-400">
+          {{ bibleVersion?.name }}
+        </div>
+      </div>
+      <div class="col">
+        <UButton
+          :icon="bibleVersion?.isDownloaded ? 'i-bx-check' : 'i-bx-download'"
+          color="primary"
+          :variant="bibleVersion?.isDownloaded ? 'ghost' : 'outline'"
+          :disabled="bibleVersion?.isDownloaded"
+          @click="downloadBibleVersion(bibleVersion?.id)"
+        >
+          {{ bibleVersion?.isDownloaded ? "Saved" : "Save" }}
+        </UButton>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { useAppStore } from "~/store/app"
+const appStore = useAppStore()
+const db = useIndexedDB()
+
+const bibleDownloadProgress = ref<string>("0")
+const bibleVersion = ref(appStore.settings.defaultBibleVersion)
+const bibleVersionLoading = ref<boolean | string>(false)
+const { bibleVersions } = storeToRefs(appStore)
+const bibleVersionOptions = ref<Array<any>>(bibleVersions.value)
+const bibleVersionSelectOptions = computed(() =>
+  [
+    ...bibleVersions.value,
+    bibleVersions.value?.find((version) => !version?.isDownloaded)
+      ? {
+          id: "+ More Versions",
+          name: "Add more versions",
+          isDownloaded: false,
+        }
+      : null,
+  ]
+    ?.filter((version) => version?.isDownloaded)
+    ?.map((version) => version?.id)
+)
+
+const isBibleVersionDownloaded = async (bibleVersion: string) => {
+  return (await db.bibleAndHymns.where("id").equals(bibleVersion).count()) > 0
+}
+
+const populateBibleVersionOptions = async () => {
+  const tempBibleVersions = [...bibleVersions.value]
+  for (const bibleVersion of tempBibleVersions) {
+    bibleVersion.isDownloaded = await isBibleVersionDownloaded(bibleVersion.id)
+  }
+  bibleVersionOptions.value = tempBibleVersions
+}
+
+const downloadBibleVersion = async (bibleVersion: string) => {
+  const tempBibleVersion = (version: string, data: any) => ({
+    id: version,
+    data,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  })
+  bibleVersionLoading.value = bibleVersion
+  const bibleResponse = await useS3File(
+    `${bibleVersion?.toLowerCase()}.json`,
+    bibleDownloadProgress
+  )
+  await db.bibleAndHymns.add(tempBibleVersion(bibleVersion, bibleResponse))
+  bibleVersionLoading.value = false
+  populateBibleVersionOptions()
+}
+
+// Function Invocations
+populateBibleVersionOptions()
+</script>
